+++ src/components/utils/index.ts
@@ -1,5 +1,5 @@
-export * from './checkIsValidNumber';
 export * from './cleanValue';
 export * from './fixedDecimalValue';
 export * from './formatValue';
+export * from './isNumber';
 export * from './padTrimValue';